+++ examples/helloworld/tests.py
@@ -1,9 +1,8 @@
 '''Tests the "helloworld" example.'''
-import unittest as test
-
 from pulsar import SERVER_SOFTWARE
 from pulsar.net import HttpClient
 from pulsar.apps.test import test_server
+from pulsar.utils.test import test
 
 from .manage import server