+++ pulsar/net/client/std.py
@@ -149,6 +149,11 @@ class HttpClientHandler(object):
 object.'''
         return self.request(url, method = 'GET')
     
+    def post(self, url, body = None):
+        if body:
+            body = to_bytestring(body)
+        return self.request(url, data = body, method = 'POST')
+    
     
 class HttpClient1(HttpClientHandler):
     '''Http handler from the standard library'''
@@ -169,8 +174,6 @@ class HttpClient1(HttpClientHandler):
         return self._opener.addheaders
     
     def request(self, url, body=None, **kwargs):
-        if body:
-            body = to_bytestring(body)
         try:
             response = self._opener.open(url,data=body,timeout=self.timeout)
         except (HTTPError,URLError) as why: